refactor utils into better categories
