+++ preliz/tests/test_distributions.py
@@ -3,9 +3,7 @@
 import pytest
 from numpy.testing import assert_almost_equal
 import numpy as np
-
 from test_helper import run_notebook
-
 
 from preliz.distributions import (
     AsymmetricLaplace,
@@ -219,6 +217,6 @@ def test_ppf(a_few_poissons):
 def test_plot_interactive(capsys, a_few_poissons):
     d_0, _, _ = a_few_poissons
     d_0.plot_interactive()
-    out, _ = capsys.readouterr()
-    assert "RuntimeError" in out
+    captured = capsys.readouterr()
+    assert "RuntimeError" in captured.out
     run_notebook("plot_interactive.ipynb")
